large-scale error corrections (primarily of a single type) and a few additions
This commit primarily replaces all previous instances of "/*" with "/$all" at the end of a line because I just discovered that unlike other adblock software, uBlock Origin doesn't honor the asterisk as a wildcard at the end of a pathname of a URL unless the pathname is just the root, i.e. initial forward slash.
The "all" keyword blocks all types of data that follow the forward-slash.
Because it was easier to do, I replaced all instances of "/*" at the end of a line with "/$all" since it achieves the same results where "/*" was already correct, i.e. after a hostname like "||heise02.webtrekk.net/*".

If you previously downloaded this list, you should delete it and replace it with this fixed version.
If you previously added this list to your uBlock Origin custom filter list, simply update to this fixed version by clicking the clock icon at the end of this filter's line and then click the "update now" button at the top of the page/tab.