DropGalaxy downloads now work and a few other additions
DropGalaxy downloads are now working again.  I am skeptical that the changes I made re-enabled downloads since the new "downloadadhash" value is not being added to the post form as an input tag, but "if it ain't still broke, don't fix it."
hurry.stream embedded videos still do not play, but it looks like it may be a problem with hurry.stream as it also has CORS blocking the m3u8 file requests on Chromium-based browsers and Firefox.  Streaming a video from hurry.stream has only ever worked for me on Android, but that was without any adblocking so have not yet tested with my adblocking rules on Android.
https://freestreams-live1.tv/nflnetwork5/ is still broken.  It appears the script which creates the embedded video iframe is being blocked at least in part.  freestreams-live1.tv works with other channels/streams though.