correct a different repeated syntax error in injected stylesheet rules
The asshole that runs Devuploads has broken downloads again.  Attempts to rectify the rules to correct this have failed so far.  Hopefully the next commit will have Devuploads working again.
The Spotify web client is back to having ads, and the playback element disappears when the ads play and doesn't return unless/until the play button is pressed for a new song/track.  So, this looks like the best workaround for now.
hurry.stream embedded videos still do not play, but it looks like it may be a problem with hurry.stream as it also has CORS blocking the m3u8 file requests on Chromium-based browsers and Firefox.  Streaming a video from hurry.stream has only ever worked for me on Android, but that was without any adblocking so have not yet tested with my adblocking rules on Android.
https://freestreams-live1.tv/nflnetwork5/ is still broken.  It appears the script which creates the embedded video iframe is being blocked at least in part.  freestreams-live1.tv works with other channels/streams though.