misc additions and removal of duplicate rules that somehow escaped detection before now
The asshole that runs Devuploads has broken downloads again as has the asshole who runs DropGalaxy.  Both file hosts appear to be using the same anti-adblock technique, but getting past it hasn't worked so far because uBlock is apparently ignoring the rules I have tried to bypass it.  I may have to remove rules or make exceptions to let more ad crap through for Devuploads and DropGalaxy to work.
The Spotify web client no longer has ads.  The workaround of having to press the play button on an individual song/track to continue playback when an ad would play does appear to be necessary sometimes.
hurry.stream embedded videos still do not play, but it looks like it may be a problem with hurry.stream as it also has CORS blocking the m3u8 file requests on Chromium-based browsers and Firefox.  Streaming a video from hurry.stream has only ever worked for me on Android, but that was without any adblocking so have not yet tested with my adblocking rules on Android.
https://freestreams-live1.tv/nflnetwork5/ is still broken.  It appears the script which creates the embedded video iframe is being blocked at least in part.  freestreams-live1.tv works with other channels/streams though.