mostly additions of garbage domains used by Android app download sites along with additions for the download sites themselves; added full screen icon/functionality to all ESPN video player frames and removed share icon to make room for it
These notes are to remind me of long-standing issues that need fixing and don't pertain to the current commit.

vidmoly still doesn't play videos, but there are no ads or ad redirects and popups.  Choose another video host/server on embed websites like kimcartoon.

The cartoon gallery @ foxnews.com is still broken.

The image carousel/gallery widget is still broken @ windowscentral.com.

Downloads @ drive.online-fix.me may still be broken, but at least they don't appear to be broken as the dialog box and buttons to trigger them is displayed.  I have not tested downloads because they require giving access to one's own Google Drive account which is an idiotic thing to do given the hostility of the person who runs online-fix.me.

gplinks.co/gplinks.in is broken as adblocking is detected.

The asshole that runs Devuploads / DropGalaxy has broken downloads again.  Both file host domains appear to be using the same anti-adblock technique, but getting past it hasn't worked so far because uBlock is apparently ignoring the rules I have tried to bypass it.  I may have to remove rules or make exceptions to let more ad crap through for Devuploads and DropGalaxy to work.

hurry.stream embedded videos still do not play, but it looks like it may be a problem with hurry.stream as it also has CORS blocking the m3u8 file requests on Chromium-based browsers and Firefox.  Streaming a video from hurry.stream has only ever worked for me on Android, but that was without any adblocking so have not yet tested with my adblocking rules on Android.

https://freestreams-live1.tv/nflnetwork5/ is still broken.  It appears the script which creates the embedded video iframe is being blocked at least in part.  freestreams-live1.tv works with other channels/streams though.