addition to fix the video player on nypost.com
These notes are to remind me of long-standing issues that need fixing and don't pertain to the current commit.

The "sign in" menu on the gog.com website does not appear when this rule set is used (on Firefox only?) , although I have not been able to isolate why that is as removing/bypassing all rules that were being applied to gog.com doesn't prevent this problem.

These rules make extensive use of the "has-text()", upward(), and remove() filters and are therefore not suitable for use with the Brave web browser's native adblocking functionality since it currently doesn't support these filters.

The embedded video player @ stapadblockuser.xyz/streamtape.com/streamadblocker.cc doesn't quite work, but the many ads and popups are gone.  The real video direct URL(s) can be found at the bottom of the modified page though, and those URL's will work with a web browser's native video player.

Free downloads no longer work @ mdiaload.com/fast-down.com because it is now using a paywall service which doesn't actually allow free downloads even when these rules are disabled.

Downloads from upfiles.com/efhjd.com do not currently work because the captcha is never displayed due to its integration with an anti-adblocking scheme.  I will try to fix this in a future update to these rules.

Playing videos from kokostream.net is broken due to the complex anti-adblocking scheme it uses, however the domain may now be defunct anyway as it redirects to a different domain which is a Chinese website.

Video streams from CricHD domains work except for streams which use escast.xyz to host the iframes since escast.xyz is defunct/offline.)

These rules do not overcome the geolocation blocks that www.ustream.to has in place for all(?) channels.  I don't believe that www.ustream.to serves the content it claims because the website only serves the free Blender videos to me regardless of the channel selected and even when uBlock Origin and my custom hosts file are disabled at least from an IP address from a US-based Internet provider.

Embedded elements on www.videogameschronicles.com and affiliated websites which require certain cookies or local storage settings don't work.  Need to manually set these values as manually setting javascript variables for this doesn't work.

The cartoon and other galleries @ foxnews.com now appear to work at least if the page is reloaded after the initial load which changes the gallery into a standard webpage with larger versions of the images to scroll through.

The image carousel/gallery widget is still broken @ windowscentral.com.

Primis videos do not work @ www.eurogamer.net.

"recommended videos" aren't displayed or played on www.tomshardware.com, www.pcgamer.com, www.laptopmag.com, www.techradar.com, www.androidcentral.com, and www.windowscentral.com.  Refreshing the page after the initial load allows the video player to initialize.

Downloads @ drive.online-fix.me are broken.  Disabling uBlock and/or Brave's shields fixes this, but downloads require giving access to one's own Google Drive account which is an idiotic thing to do given the hostility of the person who runs online-fix.me.

gplinks.co/gplinks.in is broken as adblocking is detected although the notice is suppressed/removed.

Downloads from DropGalaxy are no longer working.  Neither do downloads from Devuploads and just loading the website with ublock Origin and these rules enabled will cause Firefox to leak memory until it crashes Windows.

Downloads @ jojoy.io now work thanks to my rules changing the downloadDetails() function, but I still don't really trust that website as it may be serving up fake mod Android downloads.  The downloaded APK's aren't actually hosted by jojoy.io.

Downloads from apkmody.io are either not working properly or the site now only downloads its own downloader Android app.

The rules for PressReader domains include access tokens and account information which allow access to issues of non-free publications I have personally unlocked.  If you have a PressReader subscription or otherwise want to use a login account with PressReader, comment out the rules which contain access tokens and user account names and numbers as they will override tokens assigned to another account.  The access tokens will likely have to be updated frequently although I am uncertain of the expiration time for existing unlocked content.  Unlocking new content appears to be available only two hours after the tokens were last refreshed so if you can't unlock something, it is because it is outside of that time window and I have not previously unlocked it.

To view a newspaper or magazine you have not purchased in PressReader, simply select it and then click on the "preview in page view" or "preview in text view" button.  All pages of the magazine will be viewable if I have previously unlocked it, and it may be viewable anyway if the token I am using has been renewed or if the magazine is one of the relatively few free ones on PressReader.
If you receive a "PressReader is temporarily unavailable." error page while navigating the PressReader website, simply go back to the homepage, and search for the periodical you want and follow the link to it.
Some magazines on PressReader can't have their pages printed even if the icon is not dimmed.  The print icon is always undimmed due to a rule in this rule set.  It may be possible to reassign an event to the disabled print button, so the rule won't be removed for now.  In any case I have enabled printing for these magazines using the standard menus in the web app's UI, so use that method to print pages instead.  Regarding printing, the pages are watermarked heavily and include the user name which printed the pages so don't you dare distribute the printed pages without removing the watermark (assuming it can be done) or my account will be banned, and you will lose your unlocked access to PressReader.