miscellaneous additions and corrections including for videogameschronicles.com and frdl.to
The following are notes to remind me of outstanding issues with the functionality
of the rules in this filter list.

These rules make extensive use of the "has-text()", upward(), and remove()
filters and are therefore not suitable for use with the Brave web browser's
native adblocking functionality since it currently doesn't support these filters.

Change rules which utilize trusted-set-attr to set-attr where the values are
appropriate since trusted-set-attr rules in unauthorized third-party filter lists
like this one aren't applied.


Currrent Issues:

The spot.im chat widget does not appear on third-party websites which embed it.

The Spotify web player hangs on ads and the page must be refreshed/reloaded to be
able to play music again.
Reloading the page clears the queue sometimes.
Removing at least some of the filters fixes it although a popup appears to "refresh" the page, but
it does not empty the queue like a reload of the page does.

Download from infidrive.net don't work.  The download content node is removed
after creating/loading the node.

The embedded video player does not work @ cbsnews.com.

Free downloads no longer work @ mdiaload.com/fast-down.com because it is now
using a paywall service which doesn't actually allow free downloads even when
 these rules are disabled.

The image carousel/gallery widget is broken @ www.windowscentral.com.

These rules do not overcome the geolocation blocks that www.ustream.to has in
place for all(?) channels.  I don't believe that www.ustream.to serves the content
it claims because the website only serves the free Blender videos to me regardless
of the channel selected and even when uBlock Origin and my custom hosts file are
disabled at least from an IP address from a US-based Internet provider.

Downloads @ drive.online-fix.me are broken.  Disabling uBlock and/or Brave's
shields fixes this, but downloads require giving access to one's own Google
Drive account which is an idiotic thing to do given the hostility of the person
who runs online-fix.me.

gplinks.co/gplinks.in is broken as adblocking is detected although the notice is
suppressed/removed.

Downloads from DropGalaxy are no longer working.  Neither do downloads from
Devuploads and just loading the website with ublock Origin and these rules enabled
will cause Firefox to leak memory until it crashes Windows.

Downloads from apkmody.io are either not working properly or the site now only downloads its own downloader Android app.


Fixed Previous Issues:

Primis videos now work @ www.eurogamer.net.

Primis videos on www.videogameschronicle.com and affiliated websites now work.

"recommended videos" aren't displayed or played on www.tomshardware.com, but
they are playable on affiliated websites like some of the following:
www.windowscentral.com, www.pcgamer.com, www.laptopmag.com, www.techradar.com,
and www.androidcentral.com.
Refreshing the page after the initial load allows the video player to initialize
if it is solid black on the first load of the page.

The cartoon and other galleries @ foxnews.com now appear to work at least if the page is reloaded after the initial load which changes the gallery into a standard webpage with larger versions of the images to scroll through.

The complex anti-adblocking scheme @ kokostream.net may now be bypassed, but
playing videos is still broken most likely because the domain is defunct since it
redirects to a different domain which is a Chinese website.